# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MEDITRON · СМ-Клиника API
  description: Выявление клинически значимых находок и маршрутизация пациента
  version: 0.0.1
servers:
- url: /
  description: Текущий backend или nginx; /api/sim/* доступен только при ROUTES_SIM_ENABLED=true
...
paths:
...
  /api/sim/scenarios/{runId}/next:
    post:
      tags:
      - Демо
      operationId: next
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScenarioRunDto'
  /api/sim/scenarios/{name}/start:
    post:
      tags:
      - Демо
      summary: 'Запустить сценарий: сам с паузой stepDelaySeconds или manual=true
        — по шагу через /next'
      operationId: start
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioStartRequest'
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScenarioRunDto'
  /api/sim/clock/reset:
    post:
      tags:
      - Демо
      operationId: reset
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClockDto'
  /api/sim/clock/advance:
    post:
      tags:
      - Демо
      summary: Продвинуть модельное время (минуты) и сразу исполнить наступившие таймеры
      operationId: advance
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvanceRequest'
        required: true
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
...
  /api/sim/scenarios:
    get:
      tags:
      - Демо
      operationId: scenarios
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: {}
  /api/sim/scenarios/runs/{runId}:
    get:
      tags:
      - Демо
      operationId: status
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScenarioRunDto'
  /api/sim/crm/outbox:
    get:
      tags:
      - Демо
      summary: Все сообщения, переданные в заглушку CRM
      operationId: outbox
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationDto'
  /api/sim/clock:
    get:
      tags:
      - Демо
      operationId: clock
      responses:
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload Too Large
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClockDto'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    ScenarioRunDto:
      type: object
      properties:
        runId:
          type: string
        scenario:
          type: string
        title:
          type: string
        step:
          type: integer
          format: int32
        totalSteps:
          type: integer
          format: int32
        lastStep:
          type: string
        finished:
          type: boolean
        error:
          type: string
        log:
          type: array
          items:
            type: string
    ScenarioStartRequest:
      type: object
      properties:
        stepDelaySeconds:
          type: integer
          format: int32
        manual:
          type: boolean
    ClockDto:
      type: object
      properties:
        now:
          type: string
          format: date-time
        offsetMinutes:
          type: integer
          format: int64
    AdvanceRequest:
      type: object
      properties:
        minutes:
          type: integer
          format: int64
...
    NotificationDto:
      type: object
      properties:
        id:
          type: string
        routeId:
          type: string
        specialty:
          type: string
        templateCode:
          type: string
        templateTitle:
          type: string
        fullText:
          type: string
        shortText:
          type: string
        links:
          type: object
          additionalProperties:
            type: string
        buttons:
          type: array
          items:
            type: string
        manual:
          type: boolean
        sentBy:
          type: string
        sentAt:
          type: string
          format: date-time
        delivery:
          type: string
        deliveredAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        failedAt:
          type: string
          format: date-time
        bookedAfter:
          type: boolean
        patientId:
          type: string
        patientExternalId:
          type: string
        preferredChannel:
          type: string
...